ai-agent/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-openai>=0.10.0
livekit-plugins-silero>=0.7.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
